// green-store/app/lib/api/models/customer-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Store Management API
 * API for managing store operations including inventory, orders, and customers
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CustomerCreateRequest {
    'name': string;
    'phone'?: string | null;
    'email'?: string | null;
    'address'?: string | null;
}